Add alert dialog for deleting process

+++ app/src/main/java/com/pgaa/teamvelocity/ui/sprint/SprintFragment.kt
@@ -1,9 +1,12 @@
 package com.pgaa.teamvelocity.ui.sprint
 
+import android.app.Dialog
+import android.content.DialogInterface
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
@@ -26,7 +29,7 @@ class SprintFragment : Fragment(), OnSprintInteractionListener {
             ViewModelProviders.of(this).get(SprintViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_sprint, container, false)
         val fab = root.findViewById<FloatingActionButton>(R.id.fab)
-        fab.setOnClickListener{startSprintDetailFragment(null)}
+        fab.setOnClickListener { startSprintDetailFragment(null) }
 
         val recyclerView = root.findViewById<RecyclerView>(R.id.sprint_recyclerview)
 
@@ -48,15 +51,25 @@ class SprintFragment : Fragment(), OnSprintInteractionListener {
     }
 
     override fun onDeleteClick(sprint: Sprint) {
-        sprintViewModel.deleteSprint(sprint)
-        sprintViewModel.getAllSprint()
+
+        activity?.let {
+            AlertDialog.Builder(it)
+                .setTitle("Delete Sprint")
+                .setMessage("This is irreversible! Are you sure you want to delete ${sprint.name}")
+                .setNegativeButton("Cancel") { dialog, _ -> dialog.dismiss() }
+                .setPositiveButton("Yes") { dialog, _ ->
+                    sprintViewModel.deleteSprint(sprint)
+                    sprintViewModel.getAllSprint()
+                    dialog.dismiss() }
+                .create().show()
+
+        }
     }
 
     private fun startSprintDetailFragment(sprint: Sprint?) {
         val detailFragment = SprintDetailFragment.newInstance(sprint)
-        activity?.supportFragmentManager?.beginTransaction()?.
-            replace(R.id.fragment_container, detailFragment)?.
-            addToBackStack("detail_fragment")
+        activity?.supportFragmentManager?.beginTransaction()?.replace(R.id.fragment_container, detailFragment)
+            ?.addToBackStack("detail_fragment")
             ?.commit()
     }
 }
